bugfix shop card type + seed analysis filter

+++ lib/items.cl
@@ -539,10 +539,10 @@ typedef struct ShopItem {
 } shopitem;
 
 typedef struct ShopInstance {
-    int jokerRate;
-    int tarotRate;
-    int planetRate;
-    int spectralRate;
+    double jokerRate;
+    double tarotRate;
+    double planetRate;
+    double spectralRate;
 } shop;
 
 typedef enum JokerRarity {
